Workaround broken DocumentFragment in older versions of IE 11
In these old versions of IE 11, modifying a DocumentFragment while a
MutationObserver is active could leave the childNodes of the
DocumentFragment with either a `null` parentNode, or a parentNode
pointing to a new DocumentFragment wrapper that is not identical to the
current one, but shares prototypical values like `childNodes`.

This can lead to instances where `frag.firstChild.parentNode !== frag`,
but `frag.firstChild.parentNode.childNodes === frag.childNodes`.
Eventually, this new wrapper is GC'd and the parentNode will be null,
causing all sorts of SUPER FUN PROBLEMS.